Add files via upload
I took this code from Siraj Raval as he explained in a Youtube Tutorial and I wanted to modify it to recognize wheat in a field. I tried a few masks but I don`t want to guess, I would want to know which one is compatible with my picture. My guess is that I can`t get all the green leaves because some of them are more exposed to light in the pictures and white pixels are found in between.